+++ apps/io-wallet-user-func/src/infra/azure/functions/whitelisted-fiscal-code.ts
@@ -1,4 +1,4 @@
-import { IsFiscalCodeWhitelistedHandler } from "@/infra/http/handlers/is-fiscal-code-whitelisted";
+import { IsFiscalCodeWhitelistedHandler } from "@/infra/http/handlers/whitelisted-fiscal-code";
 import { httpAzureFunction } from "@pagopa/handler-kit-azure-func";
 
 export const IsFiscalCodeWhitelistedFunction = httpAzureFunction(
