+++ apps/io-wallet-user-func/src/infra/http/handlers/__test__/whitelisted-fiscal-code.spec.ts
@@ -1,11 +1,11 @@
-import { WhitelistedFiscalCodeRepository } from "@/fiscal-code";
+import { WhitelistedFiscalCodeRepository } from "@/whitelisted-fiscal-code";
 import * as H from "@pagopa/handler-kit";
 import * as L from "@pagopa/logger";
 import * as appInsights from "applicationinsights";
 import * as TE from "fp-ts/TaskEither";
 import { describe, expect, it } from "vitest";
 
-import { IsFiscalCodeWhitelistedHandler } from "../is-fiscal-code-whitelisted";
+import { IsFiscalCodeWhitelistedHandler } from "../whitelisted-fiscal-code";
 
 describe("IsFiscalCodeWhitelistedHandler", () => {
   const whitelistedFiscalCodeRepositoryThatReturnsWhitelistedFiscalCode: WhitelistedFiscalCodeRepository =
