+++ apps/io-wallet-user-func/src/infra/http/handlers/get-wallet-instance-status.ts
@@ -8,8 +8,8 @@ import { sendTelemetryException } from "io-wallet-common/infra/azure/appinsights
 import { logErrorAndReturnResponse } from "io-wallet-common/infra/http/error";
 
 import { WalletInstanceToStatusApiModel } from "../encoders/wallet-instance";
-import { requireFiscalCodeFromHeader } from "../fiscal-code";
 import { requireWalletInstanceId } from "../wallet-instance";
+import { requireFiscalCodeFromHeader } from "../whitelisted-fiscal-code";
 
 export const GetWalletInstanceStatusHandler = H.of((req: H.HttpRequest) =>
   pipe(
